Another redesign
I had the idea to make the NSFetchRequest the source of truth, monitor its properties for changes using KVO, and refetch as necessary. I had to figure out how to prevent the ObservableObject from publishing the first time it fetches results because that is within body. I also decided to remove the Phase pattern because it's not really async and use a Result type instead, this removes the empty case and only has success and failure.